[DLG-115] 모든 레이어에 운영 로그를 남긴다. (#59)
* refactor: MDC 필터 로그 추가
- 사용자 IP, 서버 이름, UUID 등 로깅 작업 추가

* refactor: toString 재정의
- 로깅을 위한 toString 재정의

* feat: 레이어를 나타내기 위한 어노테이션 추가
- 각 계층을 나타내기 위한 어노테이션 추가

* refactor: 에러 메시지 포맷 변경
- 에러 발생 시 로그 포맷 수정

* refactor: GUEST 역할 추가
- 인증되지 않은 사용자는 GUEST로 표시

* refactor: 디미터의 법칙 준수하도록 코드 수정
- 디미터의 법칙을 어기지 않도록 코드 수정

* refactor: 로깅을 위해 DailygeUser 속성 추가
- MDC 필터 로깅을 위해 DailygeUser 속성 추가

* refactor: 각 계층에 어노테이션 추가
- 계층을 구분하기 위한 어노테이션 추가

* refactor: 필드 제거
- 상위 필드로 이동

* refactor: 클래스명 변경 적용
- Register -> Create

* refactor: 운영 데이터 초기화 순서 변경
- 로그 데이터 때문에 초기화 순서가 꼬여서 이를 제어

* chore: TimeBased UUID 의존성 추가

* refactor: Liquibase 설정 제거
- yml로 제어하기 위해 빈 설정 제거

* feat: Log 설정 추가
- 각 레이어마다 로그를 남기도록 설정 변경

* refactor: Profile 설정 추가
- 테스트 시 로깅 비활성화

* refactor: MethodArgumentTypeMismatchException 예외 처리 분리
- 별도로 예외 처리

* refactor: 운영 데이터 프로필 설정 추가
- local, dev에서만 실행되도록 프로필 설정 추가

* test: 깨진 테스트 복구
- 공통 필드(NICKNAME, EMAIL) 사용해서 발생하는 테스트 깨짐 수정

* test: UserFixture 메서드 추가 및 메서드명 변경
- 인자를 통해 객체를 생성할 수 있도록 static 메서드 추가
- 메서드 명 변경

* refactor: 요청 경로, 메서드 정보 추가 포함
- 요청 경로 및 요청 메서드 정보 포함

* test: toString 재정의 테스트 수정
- toString을 Json 양식에 맞게 수정

* refactor: toString 재정의
- String.format 으로 통일

* test: CheckStyle 경로 반영
- +개행

* feat: MdcFilter 비동기 설정 추가
- 비동기 상황에서 문맥을 잘 복사하도록 설정 추가

* refactor: Task 생성 요청 클래스명 변경
- 클래스명 통일을 위해 Register -> Create로 변경

* test: SonarCloud 코드스멜 경고 반영
- 불필요한 Import, 사용하지 않는 변수 제거

* refactor: IP 리스트 리뷰 반영사항 수정
- 배열 순회로 변경

* feat: 비동기 로깅 설정 추가
- 비동기 상황에서도 로그를 사용할 수 있도록 설정 추가

* style: 빠진 final 키워드 추가
- 프로젝트 규칙에 따라 final 키워드 추가

* refactor: 병합 과정에서 발생한 conflict 해결

* test: 정적 메서드 사용, 깨진 테스트 주석처리
- 사용자 관련 테스트 위임

* refactor: 병합 전 코드 복원
- 잘못 머지한 코드 복원

+++ dailyge-api/src/main/java/project/dailyge/app/core/user/facade/UserFacade.java
@@ -2,7 +2,7 @@ package project.dailyge.app.core.user.facade;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.stereotype.Component;
+import project.dailyge.app.common.annotation.Facade;
 import project.dailyge.app.common.auth.DailygeToken;
 import project.dailyge.app.common.auth.TokenProvider;
 import project.dailyge.app.common.exception.CommonException;
@@ -11,15 +11,15 @@ import project.dailyge.app.core.user.application.UserWriteUseCase;
 import project.dailyge.app.core.user.external.oauth.GoogleOAuthManager;
 import project.dailyge.app.core.user.external.oauth.TokenManager;
 import project.dailyge.app.core.user.external.response.GoogleUserInfoResponse;
-import project.dailyge.entity.common.EventType;
+import static project.dailyge.document.common.UuidGenerator.createTimeBasedUUID;
+import static project.dailyge.entity.common.EventType.CREATE;
 import project.dailyge.entity.user.UserEvent;
+import static project.dailyge.entity.user.UserEvent.createEvent;
 import project.dailyge.entity.user.UserJpaEntity;
 
 import java.util.Optional;
 
-import static project.dailyge.document.common.UuidGenerator.createTimeBasedUUID;
-
-@Component
+@Facade
 @RequiredArgsConstructor
 public class UserFacade {
 
@@ -33,22 +33,22 @@ public class UserFacade {
     public DailygeToken login(final String code) throws CommonException {
         final GoogleUserInfoResponse response = googleOAuthManager.getUserInfo(code);
         final Optional<UserJpaEntity> findUserByEmail = userReadUseCase.findActiveUserByEmail(response.getEmail());
-        final Long userId = eventPublish(findUserByEmail, response);
+        final Long userId = publishEvent(findUserByEmail, response);
 
         final DailygeToken token = tokenProvider.createToken(userId, response.getEmail());
         tokenManager.saveRefreshToken(userId, token.refreshToken());
         return token;
     }
 
-    private Long eventPublish(
+    private Long publishEvent(
         final Optional<UserJpaEntity> findUserByEmail,
         final GoogleUserInfoResponse response
     ) {
         final Long userId = findUserByEmail.isEmpty() ? userWriteUseCase.getSequence() : findUserByEmail.get().getId();
-        final UserEvent event = UserEvent.createEvent(
+        final UserEvent event = createEvent(
             userId,
             createTimeBasedUUID(),
-            EventType.CREATE,
+            CREATE,
             response.getName(),
             response.getEmail(),
             response.getPicture()
@@ -58,6 +58,8 @@ public class UserFacade {
     }
 
     public void logout(final Long userId) {
-        tokenManager.deleteRefreshToken(userId);
+        if (userId != null) {
+            tokenManager.deleteRefreshToken(userId);
+        }
     }
 }
